Update fix_rigid_abrade.cpp
Work in progress parallelisation of the remshing scheme. Currently works for a single rigid body given that all processors own the target atom (as either a local or ghost atom) - there is a reliance currently on MPI_ALLreduce() within remesh() to remove angles during the timestep. If a processor has no reason to call remesh() then it will not be included in the reduce call and the simulation breaks. It is proposed that angles are instead only removed during end_of_step() and are just not processed during the timestep if marked for deletion. This method is already successfully employed for deleting atoms. Additionally, this will remove the need for global communication during the timestep with it instead being called only once at the end where processors would need to sync anyway,