Update fix_rigid_abrade.cpp
Remesh() is now called on processors which own a copy of the target body - even if they don't own the target atom - since re_setup_bodies_static() requires body communication. 

Since body properties now need to be accessed on processors which don't own the target atom, atom2body[i] cannot be used. Therefore, a body_dlist() is also constructed which stores the local body index corresponding to each atom in the dlist. 

The current state of the code is that as long as every processor owns a copy of the body to be remeshed, re-meshing is successful. However, when rebuilding the topology after remeshing there is a MPI_ALLreduce() call used for an error check. This will not work if only a subset of processors call remesh(). The next step is to only update the local topology lists during re-meshing - this is okay because build_topology() will formally be called at the end of the remeshing timestep when neighborlists are rebuilt.