Update Extended_body_test.cpp
Reframed the problem so that the positions in the mol file represent the true mesh vertices. Therefore the mass, volume, inertia, normals, and areas are all calculated with respect to the true particles (no changes needed). However, atoms are now pushed inwards towards the COM by their atom radius - as opposed to the mesh being pushed outwards. This solution is also more conducive to potentially reading in STL files as an input.