Update Extended_body_test.cpp
Moved the positioning of the owning atom to the centre of mass, set during resetup_bodies_static() and setup_bodies_static(), to before xcmiamges are reset and unwrapped coordinates are calculated. If this was not the case, the owning atoms' body coordinates were non-zero and their position in set_xv() was unstable.